Cleaned the code.

+++ MokkiApp2.0/MokkiApp/MokkiApp/Repositories/CategoryRepository.cs
@@ -11,14 +11,14 @@ namespace MokkiApp.Repositories
             _context = context ?? throw new ArgumentNullException(nameof(context));
         }
 
-        public async Task <List<Category>> GetAllAsync()
+        public async Task<List<Category>> GetAllAsync()
         {
             return await _context.Categories.ToListAsync();
         }
 
         public async Task<Category> GetAsync(int id)
         {
-            return await _context.Categories.FirstOrDefaultAsync(i => i.Id==id);
+            return await _context.Categories.FirstOrDefaultAsync(i => i.Id == id);
         }
 
         public Task<int> AddCategory(Category category)
@@ -27,16 +27,29 @@ namespace MokkiApp.Repositories
             return _context.SaveChangesAsync();
         }
 
-        public Task<int> DeleteCategory(Category category)
-        {   
+        public async Task<bool> DeleteCategory(int id)
+        {
+            var category = await _context.Categories.FindAsync(id);
+
+            if (category == null)
+            {
+                return false;
+            }
             _context.Categories.Remove(category);
-            return _context.SaveChangesAsync();
+            await _context.SaveChangesAsync();
+            return true;
+
         }
 
         public Task<int> UpdateCategory(Category category)
         {
             _context.Categories.Update(category);
             return _context.SaveChangesAsync();
+        }
+        public async Task<bool> CategoryExistsAsync(string categoryName)
+        {
+            var formattedName = categoryName.Trim().ToLower();
+            return await _context.Categories.AnyAsync(p => p.Name.Trim().ToLower() == formattedName);
         }
     }
 }
